+++ apibuilder4elastic/custom_flow_nodes/api-builder-plugin-axway-api-management/src/index.js
@@ -41,10 +41,10 @@ async function getPlugin(pluginConfig, options) {
 			if(!result.isValid) {
 				throw new Error(`Error checking configured API-Manager(s). ${JSON.stringify(result)}`);
 			} else {
-				options.logger.info("Connection to API-Manager(s) successfully validated.");
+				options.logger.info(result.message);
 			}
 		} else {
-			options.logger.warn("Config validation is skipped, as parameter: pluginConfig.validateConfig=true");
+			options.logger.warn("Config validation is skipped, as parameter: pluginConfig.validateConfig=false");
 		}
 	}