+++ apibuilder4elastic/custom_flow_nodes/api-builder-plugin-axway-api-management/test/test-customProperties-MappingMerge.js
@@ -14,8 +14,7 @@ describe('Merge custom properties tests', () => {
 
 	const envFilePath = path.join(__dirname, '.env');
 	if (fs.existsSync(envFilePath)) {
-		delete process.env.API_MANAGER; // Otherwise it is not overwritten
-		envLoader.config({ path: envFilePath });
+		envLoader.config({ override: true, path: envFilePath });
 	}
 	// Delete the cached module 
 	decache('../config/axway-api-utils.default.js');